requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anyascii==0.3.2
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
audioread==3.0.1
Babel==2.13.1
bangla==0.0.2
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
cachetools==5.3.2
cairocffi==1.6.1
CairoSVG==2.7.1
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==2.2.1
comm==0.2.0
confection==0.1.4
contourpy==1.2.0
coqpit==0.0.17
cryptography==41.0.7
cssselect2==0.7.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.6
dateparser==1.1.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.8.0
docopt==0.6.2
einops==0.7.0
encodec==0.1.1
executing==2.0.1
fastapi==0.105.0
fastjsonschema==2.19.0
filelock==3.13.1
Flask==3.0.0
fonttools==4.46.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.0
g2pkk==0.1.2
google-auth==2.24.0
google-auth-oauthlib==1.1.0
grpcio==1.59.3
grpclib==0.4.3
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
h2==4.1.0
hangul-romanize==0.1.0
hpack==4.0.0
httpcore==1.0.2
httpx==0.25.2
huggingface-hub==0.19.4
hyperframe==6.0.1
idna==3.6
importlib-metadata==7.0.0
inflect==7.0.0
iniconfig==2.0.0
ipykernel==6.27.1
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonlines==1.2.0
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.1
litellm==1.14.1
llvmlite==0.41.1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
modal==0.56.4311
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.4
murmurhash==1.0.10
nbclient==0.9.0
nbconvert==7.12.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==2.8.8
nltk==3.8.1
notebook==7.0.6
notebook_shim==0.2.3
num2words==0.5.13
numba==0.58.1
numpy==1.26.2
oauthlib==3.2.2
openai==1.3.9
overrides==7.4.0
packaging==23.2
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20221105
pdfplumber==0.10.3
pexpect==4.9.0
Pillow==10.1.0
platformdirs==4.1.0
pluggy==1.3.0
pooch==1.8.0
preshed==3.0.9
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.23.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydub==0.25.1
Pygments==2.17.2
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pynndescent==0.5.11
pyparsing==3.1.1
pypdfium2==4.24.0
pypinyin==0.49.0
pysbd==0.3.4
pytest==7.4.3
python-crfsuite==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.31.1
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
Send2Trash==1.8.2
sigtools==4.0.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.27.0
SudachiDict-core==20230927
SudachiPy==0.6.7
sympy==1.12
synchronicity==0.5.3
tblib==3.0.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.18.0
thinc==8.2.1
threadpoolctl==3.2.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
torch==2.1.1
torchaudio==2.1.1
tornado==6.4
tqdm==4.66.1
trainer==0.0.32
traitlets==5.14.0
transformers==4.35.2
TTS==0.21.3
typer==0.9.0
types-certifi==2021.10.8.3
types-python-dateutil==2.8.19.14
types-toml==0.10.8.7
typing_extensions==4.8.0
tzlocal==5.2
umap-learn==0.5.5
Unidecode==1.3.7
uri-template==1.3.0
urllib3==2.1.0
wasabi==1.1.2
watchfiles==0.21.0
wcwidth==0.2.12
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
yarl==1.9.3
zipp==3.17.0
